start sharing code between the plugins
