+++ src/plugins/FFGLTouchEngine/CMakeLists.txt
@@ -4,6 +4,8 @@ target_sources(FFGLTouchEngine PRIVATE
     TouchEngine.cpp
     ../../lib/FFGL/FFGLSDK.h
     ../../lib/FFGL/FFGLSDK.cpp
+    ../shared/TouchEnginePluginBase.h
+    ../shared/TouchEnginePluginBase.cpp
 )
 
 if (WIN32)
@@ -35,6 +37,7 @@ endif()
 target_include_directories(FFGLTouchEngine PUBLIC
     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../../../include>
     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../../lib>
+    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../../plugins/shared>
     $<INSTALL_INTERFACE:include>
 )
 target_link_libraries(FFGLTouchEngine PUBLIC OpenGL::GL)
