+++ src/plugins/FFGLTouchEngine/TouchEngine.h
@@ -1,49 +1,14 @@
 #pragma once
 
-#ifdef _WIN32
-#include <windows.h>
-#include <d3d11_4.h>
-#pragma comment(lib, "d3d11.lib")
-#pragma comment(lib, "user32.lib")   
-#pragma comment(lib, "gdi32.lib")
-#define WIN32_LEAN_AND_MEAN
-#include <wrl.h>
-#include <shobjidl.h>
-#endif
-
-#ifdef __APPLE__
-#define NS_PRIVATE_IMPLEMENTATION
-#define CA_PRIVATE_IMPLEMENTATION
-#define MTL_PRIVATE_IMPLEMENTATION
-#include <Foundation/Foundation.hpp>
-#include <Metal/Metal.hpp>
-#include <QuartzCore/QuartzCore.hpp>
-#endif
-
-#include "FFGL/FFGLSDK.h"
-#include <map>
-#include "TouchEngine/TouchObject.h"
-#include "TouchEngine/TEGraphicsContext.h"
-
-#ifdef _WIN32
-#include "TouchEngine/TED3D11.h"
-#include "SpoutGL/SpoutSender.h"
-#endif
+#include <TouchEnginePluginBase.h>
 
 #include "Thumbnail.h"
 
-
-typedef struct {
-	std::string identifier;
-	uint8_t count;
-	FFUInt32 children[4];
-} VectorParameterInfo;
-
-class FFGLTouchEngine : public CFFGLPlugin
+class FFGLTouchEngine : public FFGLTouchEnginePluginBase
 {
 public:
 	FFGLTouchEngine();
-	~FFGLTouchEngine();
+	~FFGLTouchEngine() override;
 
 	//CFFGLPlugin
 	FFResult InitGL(const FFGLViewportStruct* vp) override;
@@ -57,32 +22,9 @@ public:
 	char* GetTextParameter(unsigned int index) override;
 
 private:
-	TouchObject<TEInstance> instance;
-
-#ifdef _WIN32
-	Microsoft::WRL::ComPtr<ID3D11Device> D3DDevice;
-	TouchObject<TED3D11Context> D3DContext;
-	Microsoft::WRL::ComPtr <ID3D11Texture2D> D3DTextureInput = nullptr;
-	Microsoft::WRL::ComPtr <ID3D11Texture2D> D3DTextureOutput = nullptr;
-	std::map<ID3D11Texture2D*, IDXGIKeyedMutex*> TextureMutexMap;
-
-	DXGI_FORMAT DXFormat = DXGI_FORMAT_B8G8R8A8_UNORM;
-#endif
 #ifdef __APPLE__
     MTL::Device *pDevice;
 #endif
-	GLint GLFormat = 0;
-
-	std::atomic_bool isTouchEngineLoaded = false;
-	std::atomic_bool isTouchEngineReady = false;
-	std::atomic_bool isGraphicsContextLoaded = false;
-	std::atomic_bool isTouchFrameBusy = false;
-	uint64_t FrameCount = 0;
-
-	//Touch file capabilities
-	bool hasVideoInput = false;
-	bool hasVideoOutput = false;
-	bool isVideoFX = false;
 
 	//TouchEngine IO objects
 #ifdef _WIN32
@@ -93,23 +35,6 @@ private:
 	TouchObject<TETexture> TEVideoOutputTexture;
 	TouchObject<TEFloatBuffer> TEAudioInFloatBuffer1;
 	TouchObject<TEFloatBuffer> TEAudioInFloatBuffer2;
-	
-	//TouchEngine parameters
-	uint32_t MaxParamsByType = 0;
-	uint32_t OffsetParamsByType = 0;
-	std::set<FFUInt32> ActiveParams;
-	std::vector<std::pair<std::string, FFUInt32>> Parameters;
-	std::unordered_map<FFUInt32, FFUInt32> ParameterMapType;
-	std::unordered_map<FFUInt32, int32_t> ParameterMapInt;
-	std::unordered_map<FFUInt32, double> ParameterMapFloat;
-	std::unordered_map<FFUInt32, std::string> ParameterMapString;
-	std::unordered_map<FFUInt32, bool> ParameterMapBool;
-
-	std::set<FFUInt32> ActiveVectorParams;
-	std::vector<VectorParameterInfo> VectorParameters;
-
-	//Texture Name
-	std::string OutputOpName;
 
 
 #ifdef _WIN32
@@ -119,15 +44,7 @@ private:
 	bool isInteropInitialized = false;
 #endif
 
-	int OutputWidth = 0;
-	int OutputHeight = 0;
 
-
-	std::string FilePath;
-
-
-	ffglex::FFGLShader shader;  //!< Utility to help us compile and link some shaders into a program.
-	ffglex::FFGLScreenQuad quad;//!< Utility to help us render a full screen quad.
 #ifdef _WIN32
 	GLuint SpoutTexture = 0;
 #endif
