+++ src/plugins/FFGLTouchEngineFX/TouchEngineFX.cpp
@@ -43,35 +43,6 @@ void main()
 }
 )";
 
-std::string GetSeverityString(TESeverity severity) {
-	switch (severity)
-	{
-	case TESeverityWarning:
-		return "Warning";
-	case TESeverityError:
-		return "Error";
-	default:
-		return "Unknown";
-	}
-
-}
-
-std::string GenerateRandomString(size_t length)
-{
-	auto randchar = []() -> char
-		{
-			const char charset[] =
-				"0123456789"
-				"ABCDEFGHIJKLMNOPQRSTUVWXYZ"
-				"abcdefghijklmnopqrstuvwxyz";
-			const size_t max_index = (sizeof(charset) - 1);
-			return charset[rand() % max_index];
-		};
-	std::string str(length, 0);
-	std::generate_n(str.begin(), length, randchar);
-	return str;
-}
-
 #ifdef _WIN32
 void textureCallback(TED3D11Texture* texture, TEObjectEvent event, void* info)
 {
@@ -79,65 +50,8 @@ void textureCallback(TED3D11Texture* texture, TEObjectEvent event, void* info)
 }
 #endif
 
-FFResult FailAndLog(std::string msg) {
-
-	FFGLLog::LogToHost(msg.c_str());
-	return FF_FAIL;
-}
-
-
-#ifdef _WIN32
-DXGI_FORMAT GlToDXFromat(GLint format) {
-	
-	switch (format) {
-	case GL_RGBA8:
-		return DXGI_FORMAT_B8G8R8A8_UNORM;
-
-	case GL_RGB8:
-		return DXGI_FORMAT_R8G8B8A8_UNORM;
-
-	case GL_RGBA16:
-		return DXGI_FORMAT_R16G16B16A16_UNORM;
-
-
-
-	default:
-		auto s = "Unsupported Format:: " + std::to_string(format);
-		FFGLLog::LogToHost(s.c_str());
-	}
-}
-#endif
-
-GLenum GetGlType(GLint format) {
-	switch (format) {
-	case GL_UNSIGNED_BYTE:
-		return GL_RGBA;
-	case GL_RGBA16:
-		return GL_UNSIGNED_SHORT;
-	}
-}
-
-#ifdef _WIN32
-GLenum GetGlType(DXGI_FORMAT format) {
-	switch (format) {
-	case DXGI_FORMAT_B8G8R8A8_UNORM:
-		return GL_UNSIGNED_BYTE;
-	case DXGI_FORMAT_R16G16B16A16_UNORM:
-		return GL_UNSIGNED_SHORT;
-	case DXGI_FORMAT_R32G32B32A32_FLOAT:
-		return GL_FLOAT;
-	default:
-		return GL_UNSIGNED_BYTE;
-	}
-}
-#endif
-
 FFGLTouchEngineFX::FFGLTouchEngineFX()
-	: CFFGLPlugin(),
-isTouchEngineLoaded(false),
-isTouchEngineReady(false),
-isGraphicsContextLoaded(false),
-isTouchFrameBusy(false)
+	: FFGLTouchEnginePluginBase()
 {
 	srand(static_cast<long int>(time(0)));
 
@@ -667,7 +581,7 @@ FFResult FFGLTouchEngineFX::SetFloatParameter(unsigned int dwIndex, float value)
 		return FF_SUCCESS;
 	}
 
-	if(type == FF_TYPE_OPTION) {
+	if (type == FF_TYPE_OPTION) {
 		ParameterMapInt[dwIndex] = static_cast<int32_t>(value);
 		return FF_SUCCESS;
 	}
