+++ src/plugins/shared/TouchEnginePluginBase.cpp
@@ -1,0 +1,97 @@
+#include "TouchEnginePluginBase.h"
+
+FFResult FailAndLog(std::string message)
+{
+	FFGLLog::LogToHost(message.c_str());
+	return FF_FAIL;
+}
+
+std::string GetSeverityString(TESeverity severity) {
+	switch (severity) {
+	case TESeverityWarning:
+		return "Warning";
+	case TESeverityError:
+		return "Error";
+	default:
+		return "Unknown";
+	}
+
+}
+
+std::string GenerateRandomString(size_t length) {
+	auto randchar = []() -> char {
+		const char charset[] =
+			"0123456789"
+			"ABCDEFGHIJKLMNOPQRSTUVWXYZ"
+			"abcdefghijklmnopqrstuvwxyz";
+		const size_t max_index = (sizeof(charset) - 1);
+		return charset[rand() % max_index];
+		};
+	std::string str(length, 0);
+	std::generate_n(str.begin(), length, randchar);
+	return str;
+}
+
+#ifdef _WIN32
+DXGI_FORMAT GlToDXFromat(GLint format) {
+
+	switch (format) {
+	case GL_RGBA8:
+		return DXGI_FORMAT_B8G8R8A8_UNORM;
+
+	case GL_RGB8:
+		return DXGI_FORMAT_R8G8B8A8_UNORM;
+
+	case GL_RGBA16:
+		return DXGI_FORMAT_R16G16B16A16_UNORM;
+
+
+
+	default:
+		auto s = "Unsupported Format:: " + std::to_string(format);
+		FFGLLog::LogToHost(s.c_str());
+	}
+}
+#endif
+
+GLenum GetGlType(GLint format) {
+	switch (format) {
+	case GL_UNSIGNED_BYTE:
+		return GL_RGBA;
+	case GL_RGBA16:
+		return GL_UNSIGNED_SHORT;
+	}
+}
+
+#ifdef _WIN32
+GLenum GetGlType(DXGI_FORMAT format) {
+	switch (format) {
+	case DXGI_FORMAT_B8G8R8A8_UNORM:
+		return GL_UNSIGNED_BYTE;
+	case DXGI_FORMAT_R16G16B16A16_UNORM:
+		return GL_UNSIGNED_SHORT;
+	case DXGI_FORMAT_R32G32B32A32_FLOAT:
+		return GL_FLOAT;
+	default:
+		return GL_UNSIGNED_BYTE;
+	}
+}
+#endif
+
+FFGLTouchEnginePluginBase::FFGLTouchEnginePluginBase()
+	: CFFGLPlugin(),
+	isTouchEngineLoaded(false),
+	isTouchEngineReady(false),
+	isGraphicsContextLoaded(false),
+	isTouchFrameBusy(false)
+{
+}
+
+FFGLTouchEnginePluginBase::~FFGLTouchEnginePluginBase()
+{
+}
+
+FFResult FFGLTouchEnginePluginBase::DeInitGL()
+{
+	return FF_SUCCESS;
+}
